requirements.txt
requests
dataclasses
cozepy
mysql-connector-python
serverchan-sdk
fastapi
uvicorn
apscheduler
python-dotenv